fix: golint
