+++ plugin/servicerouter/lane/lane_router.go
@@ -29,13 +29,8 @@ package lane
 import (
 	"math"
 	"math/rand"
-	"regexp"
-	"sort"
-	"strings"
-	"sync"
 	"time"
 
-	apimodel "github.com/polarismesh/specification/source/go/api/v1/model"
 	apitraffic "github.com/polarismesh/specification/source/go/api/v1/traffic_manage"
 
 	"github.com/polarismesh/polaris-go/pkg/config"
@@ -64,348 +59,6 @@ const (
 	// serviceEntryType 服务类型流量入口的 type 值
 	serviceEntryType = "polarismesh.cn/service"
 )
-
-// regexCache 缓存已编译的正则表达式，避免高频路由路径上重复编译
-var regexCache sync.Map // map[string]*regexp.Regexp
-
-// getCompiledRegex 获取编译后的正则，命中缓存直接返回
-func getCompiledRegex(pattern string) (*regexp.Regexp, error) {
-	if v, ok := regexCache.Load(pattern); ok {
-		return v.(*regexp.Regexp), nil
-	}
-	re, err := regexp.Compile(pattern)
-	if err != nil {
-		return nil, err
-	}
-	actual, _ := regexCache.LoadOrStore(pattern, re)
-	return actual.(*regexp.Regexp), nil
-}
-
-// laneRuleItem 用于排序的泳道规则条目
-type laneRuleItem struct {
-	group      *apitraffic.LaneGroup
-	rule       *apitraffic.LaneRule
-	isEntry    bool
-	stainLabel string
-}
-
-// laneRuleContainer 泳道规则容器，维护排序后的规则列表以及快速查找索引
-type laneRuleContainer struct {
-	// sortedItems 按优先级排序后的规则条目（流量入口优先）
-	sortedItems []*laneRuleItem
-	// stainLabelIndex stainLabel -> laneRuleItem 快速查找
-	stainLabelIndex map[string]*laneRuleItem
-}
-
-// newLaneRuleContainer 创建泳道规则容器
-// 排序规则：流量入口(isEntry=true) 优先 → priority 升序 → ctime 升序
-func newLaneRuleContainer(groups []*apitraffic.LaneGroup, sourceService model.ServiceMetadata) *laneRuleContainer {
-	c := &laneRuleContainer{
-		stainLabelIndex: make(map[string]*laneRuleItem),
-	}
-	for _, group := range groups {
-		for _, rule := range group.GetRules() {
-			if !rule.GetEnable() {
-				continue
-			}
-			item := &laneRuleItem{
-				group:      group,
-				rule:       rule,
-				isEntry:    isTrafficEntry(group.GetEntries(), sourceService),
-				stainLabel: buildStainLabel(rule),
-			}
-			c.sortedItems = append(c.sortedItems, item)
-			c.stainLabelIndex[item.stainLabel] = item
-		}
-	}
-	sort.SliceStable(c.sortedItems, func(i, j int) bool {
-		a, b := c.sortedItems[i], c.sortedItems[j]
-		if a.isEntry != b.isEntry {
-			return a.isEntry
-		}
-		if a.rule.GetPriority() != b.rule.GetPriority() {
-			return a.rule.GetPriority() < b.rule.GetPriority()
-		}
-		return a.rule.GetCtime() < b.rule.GetCtime()
-	})
-	return c
-}
-
-// matchByStainLabel 根据染色标签直接查找规则。
-// 支持两种格式：
-//  1. 完整格式 "{groupName}/{ruleName}" — 优先在 stainLabelIndex 中精确匹配
-//  2. 短格式 "{laneValue}" — 当精确匹配失败时，回退到匹配 rule.DefaultLabelValue
-//  3. 歧义消解：如果存在多个匹配的规则，优先返回与 destService 相关的规则，其次返回第一个匹配的规则
-func (c *laneRuleContainer) matchByStainLabel(
-	stainLabel string,
-	sourceService, destService model.ServiceMetadata,
-) *laneRuleItem {
-	// 1. 精确匹配（完整格式）
-	if item := c.stainLabelIndex[stainLabel]; item != nil {
-		return item
-	}
-	// 2. 短格式回退 + 歧义消解
-	var firstFallback, firstRelevant *laneRuleItem
-	for _, item := range c.sortedItems {
-		if item.rule.GetDefaultLabelValue() != stainLabel {
-			continue
-		}
-		if firstFallback == nil {
-			firstFallback = item
-		}
-		// destService 必须在该泳道组的 destinations 中，否则本组与当前调用无关。
-		if destService != nil && !checkServiceInLane(item.group, destService) {
-			continue
-		}
-		if firstRelevant == nil {
-			firstRelevant = item
-		}
-		if matchTrafficRule(item.rule.GetTrafficMatchRule(), sourceService) {
-			return item
-		}
-	}
-	if firstRelevant != nil {
-		return firstRelevant
-	}
-	return firstFallback
-}
-
-// matchByRouteInfo 根据路由信息匹配规则（流量识别阶段）
-// 只对流量入口（isEntry=true）的规则做 TrafficMatchRule 匹配，
-// 非入口服务不应做首次流量识别，只响应上游已透传的染色标签。
-func (c *laneRuleContainer) matchByRouteInfo(sourceService model.ServiceMetadata) *laneRuleItem {
-	for _, item := range c.sortedItems {
-		if !item.isEntry {
-			continue
-		}
-		if matchTrafficRule(item.rule.GetTrafficMatchRule(), sourceService) {
-			return item
-		}
-	}
-	return nil
-}
-
-// buildStainLabel 构建染色标签：{groupName}/{ruleName}
-func buildStainLabel(rule *apitraffic.LaneRule) string {
-	return rule.GetGroupName() + "/" + rule.GetName()
-}
-
-// isTrafficEntry 判断当前调用方是否为该泳道组的流量入口
-func isTrafficEntry(entries []*apitraffic.TrafficEntry, sourceService model.ServiceMetadata) bool {
-	if sourceService == nil {
-		return false
-	}
-	routeLogger := log.GetRouteLogger()
-	debugEnabled := routeLogger.IsLevelEnabled(log.DebugLog)
-	for _, entry := range entries {
-		switch entry.GetType() {
-		case gatewayEntryType:
-			sel := &apitraffic.ServiceGatewaySelector{}
-			if err := entry.GetSelector().UnmarshalTo(sel); err != nil {
-				if debugEnabled {
-					routeLogger.Debugf("[Router][Lane] isTrafficEntry: failed to unmarshal gateway selector, err=%v", err)
-				}
-				continue
-			}
-			svcMatched := matchSelectorService(sel.GetNamespace(), sel.GetService(), sourceService)
-			labelMatched := matchSelectorLabels(sel.GetLabels(), sourceService.GetMetadata())
-			if debugEnabled {
-				routeLogger.Debugf("[Router][Lane] isTrafficEntry: gateway entry check, "+
-					"selector=%s/%s, source=%s/%s, svcMatched=%v, labelMatched=%v",
-					sel.GetNamespace(), sel.GetService(),
-					sourceService.GetNamespace(), sourceService.GetService(),
-					svcMatched, labelMatched)
-			}
-			if svcMatched && labelMatched {
-				return true
-			}
-		case serviceEntryType:
-			sel := &apitraffic.ServiceSelector{}
-			if err := entry.GetSelector().UnmarshalTo(sel); err != nil {
-				if debugEnabled {
-					routeLogger.Debugf("[Router][Lane] isTrafficEntry: failed to unmarshal service selector, err=%v", err)
-				}
-				continue
-			}
-			svcMatched := matchSelectorService(sel.GetNamespace(), sel.GetService(), sourceService)
-			labelMatched := matchSelectorLabels(sel.GetLabels(), sourceService.GetMetadata())
-			if debugEnabled {
-				routeLogger.Debugf("[Router][Lane] isTrafficEntry: service entry check, "+
-					"selector=%s/%s, source=%s/%s, svcMatched=%v, labelMatched=%v",
-					sel.GetNamespace(), sel.GetService(),
-					sourceService.GetNamespace(), sourceService.GetService(),
-					svcMatched, labelMatched)
-			}
-			if svcMatched && labelMatched {
-				return true
-			}
-		default:
-			// 未知入口类型，跳过
-			routeLogger.Warnf("[Router][Lane] isTrafficEntry: unknown entry type %q, skipped",
-				entry.GetType())
-		}
-	}
-	return false
-}
-
-// matchSelectorService 匹配 selector 中的 namespace/service 是否和调用方一致
-func matchSelectorService(ns, svc string, sourceService model.ServiceMetadata) bool {
-	if ns != "" && ns != sourceService.GetNamespace() {
-		return false
-	}
-	if svc != "" && svc != sourceService.GetService() {
-		return false
-	}
-	return true
-}
-
-// matchSelectorLabels 匹配 selector 标签
-func matchSelectorLabels(labels map[string]*apimodel.MatchString, metadata map[string]string) bool {
-	if len(labels) == 0 {
-		return true
-	}
-	if metadata == nil {
-		return false
-	}
-	for k, matchStr := range labels {
-		v, ok := metadata[k]
-		if !ok {
-			return false
-		}
-		if !matchStringValue(matchStr, v) {
-			return false
-		}
-	}
-	return true
-}
-
-// matchTrafficRule 匹配 TrafficMatchRule。
-// 无规则或无匹配参数时返回 false（不匹配），避免空规则意外染色所有流量。
-func matchTrafficRule(rule *apitraffic.TrafficMatchRule, sourceService model.ServiceMetadata) bool {
-	if rule == nil {
-		return false
-	}
-	args := rule.GetArguments()
-	if len(args) == 0 {
-		return false
-	}
-	routeLogger := log.GetRouteLogger()
-	debugEnabled := routeLogger.IsLevelEnabled(log.DebugLog)
-	isAND := rule.GetMatchMode() == apitraffic.TrafficMatchRule_AND
-	for i, arg := range args {
-		trafficValue := findTrafficValue(arg, sourceService)
-		matched := matchStringValue(arg.GetValue(), trafficValue)
-		if debugEnabled {
-			routeLogger.Debugf("[Router][Lane] matchTrafficRule: arg[%d] type=%s, key=%q, "+
-				"expected=%q(%s), actual=%q, matched=%v, mode=%s",
-				i, arg.GetType(), arg.GetKey(),
-				arg.GetValue().GetValue().GetValue(), arg.GetValue().GetType(),
-				trafficValue, matched, rule.GetMatchMode())
-		}
-		if isAND && !matched {
-			return false
-		}
-		if !isAND && matched {
-			return true
-		}
-	}
-	// AND 模式全部匹配返回 true；OR 模式全部不匹配返回 false
-	return isAND
-}
-
-// findTrafficValue 根据 SourceMatch 类型从调用方元数据中提取流量值.
-//
-// 与 polaris-java LaneUtils.findTrafficValue (polaris-plugins/polaris-plugins-router/
-// router-lane/src/main/java/com/tencent/polaris/plugins/router/lane/LaneUtils.java) 保持
-// 相同的 7 类匹配维度: HEADER / CUSTOM / METHOD / CALLER_IP / COOKIE / QUERY / PATH,
-// 外加 Go 特有的 CALLER_METADATA (从 SourceService.Metadata 直接取 arg.Key).
-//
-// 查询策略:
-//
-//	api.go / api/consumer.go 的 convert() 已经用 Argument.ToLabels 把 Arguments
-//	摊平到 SourceService.Metadata 的带前缀 label key (如 "$header.user" / "$method"
-//	/ "$caller_ip"), 因此这里 lane router 和 rulebase 一样, 只读 SourceService.Metadata,
-//	按 SourceMatch 类型拼出同样的前缀 key 回查即可。
-//
-// 类型驱动的好处:
-//   - METHOD / CALLER_IP / PATH 三类匹配维度直接用固定 label key (arg.Key 为空), 不会相互覆盖;
-//   - HEADER / QUERY / COOKIE 依靠不同前缀天然分开, 即便同名 key (如都叫 user) 也独立.
-func findTrafficValue(arg *apitraffic.SourceMatch, sourceService model.ServiceMetadata) string {
-	// 读取 sourceService.Metadata 的指定 label key
-	lookup := func(labelKey string) string {
-		if labelKey == "" || sourceService == nil {
-			return ""
-		}
-		meta := sourceService.GetMetadata()
-		if v, ok := meta[labelKey]; ok {
-			return v
-		}
-		return ""
-	}
-
-	switch arg.GetType() {
-	case apitraffic.SourceMatch_HEADER:
-		return lookup(model.LabelKeyHeader + arg.GetKey())
-	case apitraffic.SourceMatch_QUERY:
-		return lookup(model.LabelKeyQuery + arg.GetKey())
-	case apitraffic.SourceMatch_COOKIE:
-		return lookup(model.LabelKeyCookie + arg.GetKey())
-	case apitraffic.SourceMatch_METHOD:
-		return lookup(model.LabelKeyMethod)
-	case apitraffic.SourceMatch_CALLER_IP:
-		return lookup(model.LabelKeyCallerIp)
-	case apitraffic.SourceMatch_PATH:
-		return lookup(model.LabelKeyPath)
-	case apitraffic.SourceMatch_CUSTOM:
-		// CUSTOM 维度的 Key 由用户自定义, 不套前缀 (与 Argument.ToLabels 的 Custom 分支一致).
-		return lookup(arg.GetKey())
-	case apitraffic.SourceMatch_CALLER_METADATA:
-		// CALLER_METADATA 直接读调用方的原始业务 metadata (不套前缀).
-		if sourceService != nil {
-			return sourceService.GetMetadata()[arg.GetKey()]
-		}
-		return ""
-	default:
-		// 未知类型兜底: 按原始 key 直查, 避免引入隐式默认行为.
-		return lookup(arg.GetKey())
-	}
-}
-
-// matchStringValue 匹配 MatchString 规则
-func matchStringValue(ms *apimodel.MatchString, value string) bool {
-	if ms == nil {
-		return true
-	}
-	pattern := ms.GetValue().GetValue()
-	switch ms.GetType() {
-	case apimodel.MatchString_EXACT:
-		return value == pattern
-	case apimodel.MatchString_NOT_EQUALS:
-		return value != pattern
-	case apimodel.MatchString_REGEX:
-		re, err := getCompiledRegex(pattern)
-		if err != nil {
-			return false
-		}
-		return re.MatchString(value)
-	case apimodel.MatchString_IN:
-		for _, part := range strings.Split(pattern, ",") {
-			if strings.TrimSpace(part) == value {
-				return true
-			}
-		}
-		return false
-	case apimodel.MatchString_NOT_IN:
-		for _, part := range strings.Split(pattern, ",") {
-			if strings.TrimSpace(part) == value {
-				return false
-			}
-		}
-		return true
-	default:
-		return value == pattern
-	}
-}
 
 // tryStainCurrentTraffic 对当前请求流量进行染色判断
 // 返回 true 表示本次流量应进入泳道，false 表示不进入
@@ -442,31 +95,6 @@ func (r *LaneRouter) tryStainByPercentage(pct *apitraffic.TrafficGray_Percentage
 		"[Router][Lane] percentage stain: percent=%d, roll=%d, stained=%v",
 		percent, roll, stained)
 	return stained
-}
-
-// parseWarmupEtime 解析规则启用时间字符串，返回 SDK 本地时区的 time.Time。
-//
-// Polaris 服务端返回的 etime 有两种常见格式：
-//   - RFC3339（带时区，如 "2026-04-21T19:14:30+08:00"）—— 可直接 time.Parse
-//   - 无时区的本地时间字符串（如 "2026-04-21 19:14:30"）—— 必须用 ParseInLocation
-//     以 time.Local 解析，否则 Go 默认按 UTC 解析会把 SDK 所在本地时区的时间当成 UTC，
-//     导致 uptime = now(本地) - startTime(被当 UTC) = 时差 × -1（如 CST 为 -8h），
-//     永远 <= 0，tryStainByWarmup 永不染色。
-//
-// 无法解析时返回零值，调用方应回退到 time.Now()。
-func parseWarmupEtime(etimeStr string) time.Time {
-	if etimeStr == "" {
-		return time.Time{}
-	}
-	if t, err := time.Parse(time.RFC3339, etimeStr); err == nil {
-		return t
-	}
-	for _, layout := range []string{"2006-01-02 15:04:05", "2006-01-02T15:04:05"} {
-		if t, err := time.ParseInLocation(layout, etimeStr, time.Local); err == nil {
-			return t
-		}
-	}
-	return time.Time{}
 }
 
 // tryStainByWarmup 按预热模式进行流量染色
@@ -517,76 +145,6 @@ func (r *LaneRouter) tryStainByWarmup(rule *apitraffic.LaneRule, warmup *apitraf
 			uptime, warmupInterval, c, probability, roll, stained, rule.GetName())
 	}
 	return stained
-}
-
-// checkServiceInLane 检查目标服务是否在泳道组的 destinations 中
-func checkServiceInLane(group *apitraffic.LaneGroup, destService model.ServiceMetadata) bool {
-	if destService == nil {
-		return false
-	}
-	routeLogger := log.GetRouteLogger()
-	debugEnabled := routeLogger.IsLevelEnabled(log.DebugLog)
-	dests := group.GetDestinations()
-	for i, dest := range dests {
-		destNs := dest.GetNamespace()
-		destSvc := dest.GetService()
-		if destNs != "" && destNs != destService.GetNamespace() {
-			if debugEnabled {
-				routeLogger.Debugf("[Router][Lane] checkServiceInLane: group=%s dest[%d] ns mismatch, "+
-					"ruleNs=%q, actualNs=%q", group.GetName(), i, destNs, destService.GetNamespace())
-			}
-			continue
-		}
-		if destSvc != "" && destSvc != destService.GetService() {
-			if debugEnabled {
-				routeLogger.Debugf("[Router][Lane] checkServiceInLane: group=%s dest[%d] svc mismatch, "+
-					"ruleSvc=%q, actualSvc=%q", group.GetName(), i, destSvc, destService.GetService())
-			}
-			continue
-		}
-		if debugEnabled {
-			routeLogger.Debugf("[Router][Lane] checkServiceInLane: group=%s matched dest[%d]=%s/%s",
-				group.GetName(), i, destNs, destSvc)
-		}
-		return true
-	}
-	if debugEnabled {
-		routeLogger.Debugf("[Router][Lane] checkServiceInLane: group=%s no dest matched, "+
-			"destsCount=%d, target=%s/%s",
-			group.GetName(), len(dests), destService.GetNamespace(), destService.GetService())
-	}
-	return false
-}
-
-// getLaneKey 获取泳道规则对应的实例元数据 key
-func getLaneKey(rule *apitraffic.LaneRule) string {
-	if k := rule.GetLabelKey(); k != "" {
-		return k
-	}
-	return instanceLaneKey
-}
-
-// buildEnabledLaneValues 构建所有已启用泳道规则的 defaultLabelValue 集合
-// 返回：map[laneMetaKey]set(enabledLaneValue)
-func buildEnabledLaneValues(groups []*apitraffic.LaneGroup) map[string]map[string]struct{} {
-	result := make(map[string]map[string]struct{})
-	for _, group := range groups {
-		for _, rule := range group.GetRules() {
-			if !rule.GetEnable() {
-				continue
-			}
-			laneKey := getLaneKey(rule)
-			laneVal := rule.GetDefaultLabelValue()
-			if laneVal == "" {
-				continue
-			}
-			if result[laneKey] == nil {
-				result[laneKey] = make(map[string]struct{})
-			}
-			result[laneKey][laneVal] = struct{}{}
-		}
-	}
-	return result
 }
 
 // LaneRouter 泳道路由插件
@@ -768,16 +326,8 @@ func (r *LaneRouter) GetFilteredInstances(
 	}
 	stainLabel, alreadyStained := srcMeta[model.LabelKeyHeader+trafficStainLabel]
 
-	sourceNs, sourceSvc := "", ""
-	if routeInfo.SourceService != nil {
-		sourceNs = routeInfo.SourceService.GetNamespace()
-		sourceSvc = routeInfo.SourceService.GetService()
-	}
-	destNs, destSvc := "", ""
-	if routeInfo.DestService != nil {
-		destNs = routeInfo.DestService.GetNamespace()
-		destSvc = routeInfo.DestService.GetService()
-	}
+	sourceNs, sourceSvc := extractNsSvc(routeInfo.SourceService)
+	destNs, destSvc := extractNsSvc(routeInfo.DestService)
 
 	if r.logCtx.GetRouteLogger().IsLevelEnabled(log.DebugLog) {
 		r.logCtx.GetRouteLogger().Debugf(
